Separate data classes.
