+++ src/hpotk/algorithm/similarity/__init__.py
@@ -1,2 +1,3 @@
-from ._ic import calculate_ic_for_annotated_items, AnnotationIcContainer
-from ._resnik import precalculate_resnik_similarity_for_hpo, SimilarityContainer
+from ._ic import calculate_ic_for_annotated_items
+from ._model import AnnotationIcContainer, SimpleAnnotationIcContainer, SimilarityContainer
+from ._resnik import precalculate_resnik_similarity_for_hpo
